change R_qsort definition to match what's in R_HOME/include/R_ext/Utils.h
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/cancerclass@70563 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/help.h
@@ -33,4 +33,5 @@ void writedata(double *array, int size, char *filename);
 void writeheader(char *filename);
 
 int mini(double a, double b);
-void R_qsort(double *v, int i, int j);
+void R_qsort    (double *v,         size_t i, size_t j);
+
